feat(usb-audio): add 48k/24-bit streaming alt
UAC2 streaming interface now exposes alt 1 (48 kHz / 16-bit PCM,
the original v1 format) and alt 2 (48 kHz / 24-bit packed PCM,
3-byte samples) alongside the mandatory alt 0. The host picks at
SET_INTERFACE time; the audio task parses 2- or 3-byte
little-endian samples accordingly and feeds spdif a single
24-bit-left-aligned int32 representation either way.

The spdif encoder consumes that int32 directly: audio MSB sits
at bit 31 and audio LSB at bit 8. 16-bit input is shifted left
16; 24-bit input is unpacked from three little-endian bytes.
The per-frame ring grows from int16[2] to int32[2] (~64 KB
total) - well within the RP2040's 264 KB SRAM.

+++ firmware/src/spdif.h
@@ -19,7 +19,10 @@ void spdif_start(unsigned int pin);
 void spdif_set_mode(spdif_mode_t mode);
 spdif_mode_t spdif_get_mode(void);
 const char *spdif_mode_name(spdif_mode_t mode);
-unsigned int spdif_write_pcm(const int16_t *samples, unsigned int frame_count);
+// samples is interleaved L/R 24-bit audio left-aligned in int32_t: the audio
+// MSB sits at bit 31 and the audio LSB at bit 8. Bits 7..0 are ignored. 16-bit
+// callers shift their value left by 16 before calling.
+unsigned int spdif_write_pcm(const int32_t *samples, unsigned int frame_count);
 unsigned int spdif_buffered_frames(void);
 void spdif_clear_usb_buffer(void);
 void spdif_take_usb_stats(spdif_usb_stats_t *stats);
